Move the contexts module out of utils
A newly created `tracing` sub-module (technically sub-package) will
over time contain all functionality pertaining to program capture from
Python. In this commit, the utils/contexts.py module is moved over
without code changes.
