+++ frontend/catalyst/ag_primitives.py
@@ -44,7 +44,7 @@ from tensorflow.python.autograph.pyct.origin_info import LineLocation
 
 import catalyst
 from catalyst.ag_utils import AutoGraphError
-from catalyst.utils.contexts import EvaluationContext
+from catalyst.tracing.contexts import EvaluationContext
 from catalyst.utils.jax_extras import DynamicJaxprTracer, ShapedArray
 from catalyst.utils.patching import Patcher
