+++ frontend/catalyst/pennylane_extensions.py
@@ -72,7 +72,11 @@ from catalyst.jax_tracer import (
     trace_quantum_tape,
     unify_result_types,
 )
-from catalyst.utils.contexts import EvaluationContext, EvaluationMode, JaxTracingContext
+from catalyst.tracing.contexts import (
+    EvaluationContext,
+    EvaluationMode,
+    JaxTracingContext,
+)
 from catalyst.utils.exceptions import CompileError, DifferentiableCompileError
 from catalyst.utils.jax_extras import (
     ClosedJaxpr,
