+++ frontend/catalyst/tracing/contexts.py
@@ -1,4 +1,4 @@
-# Copyright 2022-2023 Xanadu Quantum Technologies Inc.
+# Copyright 2022-2024 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 """
-Tracing module.
+This module provides context classes to manage and query Catalyst's and JAX's tracing state.
 """
 
 from contextlib import contextmanager
